2 commit - create deadlock and report
