+++ oznobishina/task8/jfr/src/main/java/org/ksu/model/Customer.java
@@ -1,4 +1,4 @@
-package org.example.model;
+package org.ksu.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
