+++ oznobishina/task8/jfr/src/main/resources/application.properties
