+++ oznobishina/task8/readme.md
@@ -18,6 +18,15 @@
 5) На основе анализа журнала JFR найти проблемы в приложении и обосовать, как прийти к такому 
 заключению о проблемах. В отчёт можно включать скрини JMC, ссылки на исходный код."
 
-В дампе потока видно что он ожидает блокировку
+
+Запускаем приложение Task8Main
+Запуск java flight recorder
 ![img.png](img.png)
+
+Во время работы JFR запускаем тест план в JMeter 'My test plan.jmx'
+![img_2.png](img_2.png)
+
+В JDK Mission Control в дампе потоков видно, что они ожидают блокировку.
+![img_1.png](img_1.png)
+Это не удивительно, ведь мы сэмитировали проблему с deadlock ))
